updated sprite handler and added a test sprite

+++ ludum_dare_23/Player.cpp
@@ -17,8 +17,8 @@ int Player::Draw(sf::RenderTarget *rt) const
 
 	sf::Sprite sp; 
 	sp.SetImage(*this->img);
-	sp.SetX(this->pos.x);
-	sp.SetY(this->pos.y);
+	sp.SetX(static_cast<float>(this->pos.x));
+	sp.SetY(static_cast<float>(this->pos.y));
 	rt->Draw(sp);
 
 	return 0;
